Simplify clean() in tags using Mixin
