+++ app/education/serializers.py
@@ -42,15 +42,11 @@ class SourceSerializer(_StreamFieldRepresentationMixin, serializers.Serializer):
             "image": self._serialize_stream_field(instance, "image"),
             "media": self._serialize_stream_field(instance, "media"),
             "youtube": self._serialize_stream_field(instance, "youtube"),
-            "media_caption": self._serialize_stream_field(
-                instance, "media_caption"
-            ),
+            "media_caption": self._serialize_stream_field(instance, "media_caption"),
             "source_featured_link": self._serialize_stream_field(
                 instance, "source_featured_link"
             ),
-            "source_featured_external_link": (
-                instance.source_featured_external_link
-            ),
+            "source_featured_external_link": (instance.source_featured_external_link),
             "description": self._serialize_stream_field(instance, "description"),
             "question": self._serialize_stream_field(instance, "question"),
         }
